Tensor datatype is automatically inferred from numpy array.

+++ cats_vs_dogs/src/pytorch_impl/src/train.py
@@ -45,8 +45,7 @@ def main(train_dir,
     for step, path, image, label in tqdm(preproc.preprocess_classes(steps, train_dir, label_dict),
                                          desc='Progress', total=steps, ncols=99, unit='image'):
         optimizer.zero_grad()
-        image, label = (torch.tensor(image, dtype=torch.float32).to(device),
-                        torch.tensor(label, dtype=torch.float32).to(device))
+        image, label = torch.tensor(image).to(device), torch.tensor(label).to(device)
         output = model(image)
         objective = torch.sqrt(torch.nn.functional.mse_loss(output, label))
         objective.backward()
